Rename octtrees -> octrees
